+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lite-ssr/core
 
+## 0.4.5
+
+### Patch Changes
+
+- Add request pass to renderer
+
 ## 0.4.4
 
 ### Patch Changes